scripts/requirements.txt
python-dotenv
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
beautifulsoup4
PyYAML
